Address timetable review feedback
