+++ src/ludamus/adapters/web/django/templatetags/tessera/button.py
@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 
-from django.utils.html import format_html
+from django.utils.html import format_html, format_html_join
 from django.utils.safestring import mark_safe
 
 from .icon import icon as render_icon
@@ -30,8 +30,7 @@ def render_button(  # ruff:ignore[too-many-arguments] — template-tag adapter; 
     disabled: bool = False,
     icon: str | None = None,
     full_width_mobile: bool | None = None,
-    onclick: str | None = None,
-    title: str | None = None,
+    **attrs: str | int | bool | None,
 ) -> str:
     """Render a styled button (``<button>``) or link button (``<a>``).
 
@@ -43,8 +42,8 @@ def render_button(  # ruff:ignore[too-many-arguments] — template-tag adapter; 
     link buttons in toolbars. Defaults to ``href is None``: form-submit
     buttons stretch, link buttons don't.
 
-    ``title`` renders a native tooltip — mainly for explaining *why* a
-    disabled button is disabled.
+    Extra keyword arguments render as escaped HTML attributes; underscores
+    become hyphens.
 
     Returns:
         HTML string of the rendered button.
@@ -71,18 +70,26 @@ def render_button(  # ruff:ignore[too-many-arguments] — template-tag adapter; 
     body = format_html(
         "{}{}", mark_safe(icon_html), text  # ruff: ignore[suspicious-mark-safe-usage]
     )
-    title_attr = format_html(' title="{}"', title) if title is not None else ""
+    rendered_attrs = format_html_join(
+        "",
+        ' {}="{}"',
+        (
+            (name.replace("_", "-"), value)
+            for name, value in attrs.items()
+            if value is not None and value is not False
+        ),
+    )
 
     if href is not None:
         if disabled:
             return format_html(
                 '<a class="{}" aria-disabled="true" tabindex="-1"{}>{}</a>',
                 class_str,
-                title_attr,
+                rendered_attrs,
                 body,
             )
         return format_html(
-            '<a href="{}" class="{}"{}>{}</a>', href, class_str, title_attr, body
+            '<a href="{}" class="{}"{}>{}</a>', href, class_str, rendered_attrs, body
         )
 
     if disabled:
@@ -90,22 +97,13 @@ def render_button(  # ruff:ignore[too-many-arguments] — template-tag adapter; 
             '<button type="{}" class="{}" disabled{}>{}</button>',
             button_type,
             class_str,
-            title_attr,
-            body,
-        )
-    if onclick is not None:
-        return format_html(
-            '<button type="{}" class="{}" onclick="{}"{}>{}</button>',
-            button_type,
-            class_str,
-            onclick,
-            title_attr,
+            rendered_attrs,
             body,
         )
     return format_html(
         '<button type="{}" class="{}"{}>{}</button>',
         button_type,
         class_str,
-        title_attr,
+        rendered_attrs,
         body,
     )
