+++ src/ludamus/gates/web/django/chronology/panel/views/timetable.py
@@ -19,17 +19,14 @@ from ludamus.gates.web.django.chronology.panel.views.base import (
     PanelAccessMixin,
     PanelRequest,
 )
-from ludamus.mills.chronology import (
-    ConflictDetectionService,
-    TimetableOverviewService,
-    TimetableService,
-)
+from ludamus.mills.chronology import ConflictDetectionService, TimetableOverviewService
+from ludamus.mills.timetable import TimetableService
 from ludamus.pacts import (
     UNSCHEDULED_LIST_LIMIT,
     NotFoundError,
     UnscheduledSessionFilter,
 )
-from ludamus.pacts.chronology import SessionPlacement
+from ludamus.pacts.chronology import DateSelection, SessionPlacement
 
 
 def _parse_iso_duration_minutes(iso: str) -> int:
@@ -58,22 +55,15 @@ def _build_back_url(slug: str, query: QueryDict) -> str:
     return f"{base}?{urlencode(params)}" if params else base
 
 
-def _parse_date_selection(raw: str | None) -> tuple[date | None, bool, str]:
+def _parse_date_selection(raw: str | None) -> DateSelection:
     if raw == "all":
-        return None, True, "all"
-    if not raw:
-        return None, False, ""
-    try:
-        selected_date = date.fromisoformat(raw)
-    except ValueError:
-        return None, False, ""
-    return selected_date, False, selected_date.isoformat()
-
-
-def _grid_date_param(grid_selected_date: date | None, *, show_all_days: bool) -> str:
-    if show_all_days:
         return "all"
-    return grid_selected_date.isoformat() if grid_selected_date else ""
+    if not raw:
+        return "all"
+    try:
+        return date.fromisoformat(raw)
+    except ValueError:
+        return "all"
 
 
 class TimetablePageView(PanelAccessMixin, EventContextMixin, View):
@@ -97,9 +87,7 @@ class TimetablePageView(PanelAccessMixin, EventContextMixin, View):
         except ValueError:
             room_page = 1
 
-        selected_date, show_all_days, _ = _parse_date_selection(
-            self.request.GET.get("date")
-        )
+        date_selection = _parse_date_selection(self.request.GET.get("date"))
 
         category_pk_raw = self.request.GET.get("category", "").strip()
         category_pk = int(category_pk_raw) if category_pk_raw.isdigit() else None
@@ -112,7 +100,7 @@ class TimetablePageView(PanelAccessMixin, EventContextMixin, View):
             tz=get_current_timezone(),
             track_pk=filter_track_pk,
             space_page=room_page,
-            date_selection="all" if show_all_days else selected_date,
+            date_selection=date_selection,
         )
         conflict_service = ConflictDetectionService(uow)
         conflicts = conflict_service.list_all_for_track(
@@ -135,10 +123,7 @@ class TimetablePageView(PanelAccessMixin, EventContextMixin, View):
         context["category_pk"] = category_pk
         context["max_duration_minutes"] = max_duration_minutes
         context["duration_chips"] = [("≤30 min", 30), ("≤60 min", 60), ("≤90 min", 90)]
-        context["selected_date"] = grid.selected_date
-        context["date_param"] = _grid_date_param(
-            grid.selected_date, show_all_days=grid.show_all_days
-        )
+        context["date_selection"] = grid.date_selection
         context["slug"] = slug
         context["tab_urls"] = _timetable_tab_urls(slug)
         context["active_tab"] = "timetable"
@@ -163,9 +148,7 @@ class TimetableSessionListPartView(PanelAccessMixin, EventContextMixin, View):
         category_pk = int(category_pk_raw) if category_pk_raw.isdigit() else None
         max_dur_raw = self.request.GET.get("max_duration", "").strip()
         max_duration_minutes = int(max_dur_raw) if max_dur_raw.isdigit() else None
-        selected_date, _, date_param = _parse_date_selection(
-            self.request.GET.get("date")
-        )
+        date_selection = _parse_date_selection(self.request.GET.get("date"))
 
         uow = self.request.di.uow
         sessions, has_more = uow.sessions.list_unscheduled_by_event(
@@ -175,7 +158,9 @@ class TimetableSessionListPartView(PanelAccessMixin, EventContextMixin, View):
                 search=search,
                 max_duration_minutes=max_duration_minutes,
                 category_pk=category_pk,
-                available_on=selected_date,
+                available_on=(
+                    date_selection if isinstance(date_selection, date) else None
+                ),
             ),
         )
         categories = uow.proposal_categories.list_by_event(current_event.pk)
@@ -192,8 +177,7 @@ class TimetableSessionListPartView(PanelAccessMixin, EventContextMixin, View):
             "max_duration_minutes": max_duration_minutes,
             "duration_chips": duration_chips,
             "filter_track_pk": filter_track_pk,
-            "selected_date": selected_date,
-            "date_param": date_param,
+            "date_selection": date_selection,
             "slug": slug,
         }
         return TemplateResponse(
@@ -218,17 +202,14 @@ class TimetableBrowsePanePartView(PanelAccessMixin, EventContextMixin, View):
         max_dur_raw = self.request.GET.get("max_duration", "").strip()
         max_duration_minutes = int(max_dur_raw) if max_dur_raw.isdigit() else None
         search = self.request.GET.get("search", "").strip()
-        selected_date, _, date_param = _parse_date_selection(
-            self.request.GET.get("date")
-        )
+        date_selection = _parse_date_selection(self.request.GET.get("date"))
 
         context = {
             "filter_track_pk": filter_track_pk,
             "category_pk": category_pk,
             "max_duration_minutes": max_duration_minutes,
             "search": search,
-            "selected_date": selected_date,
-            "date_param": date_param,
+            "date_selection": date_selection,
             "slug": slug,
             "current_event": current_event,
         }
@@ -305,9 +286,7 @@ class TimetableGridPartView(PanelAccessMixin, EventContextMixin, View):
         except ValueError:
             room_page = 1
 
-        selected_date, show_all_days, _ = _parse_date_selection(
-            self.request.GET.get("date")
-        )
+        date_selection = _parse_date_selection(self.request.GET.get("date"))
 
         uow = self.request.di.uow
         grid = TimetableService(uow).build_grid(
@@ -315,7 +294,7 @@ class TimetableGridPartView(PanelAccessMixin, EventContextMixin, View):
             tz=get_current_timezone(),
             track_pk=filter_track_pk,
             space_page=room_page,
-            date_selection="all" if show_all_days else selected_date,
+            date_selection=date_selection,
         )
         slot_violations = ConflictDetectionService(uow).list_preferred_slot_violations(
             event_pk=current_event.pk, track_pk=filter_track_pk
@@ -326,9 +305,7 @@ class TimetableGridPartView(PanelAccessMixin, EventContextMixin, View):
             "filter_track_pk": filter_track_pk,
             "conflict_session_pks": set(),
             "slot_violation_session_pks": {v.session_pk for v in slot_violations},
-            "date_param": _grid_date_param(
-                grid.selected_date, show_all_days=grid.show_all_days
-            ),
+            "date_selection": grid.date_selection,
             "slug": slug,
         }
         return TemplateResponse(
@@ -401,7 +378,9 @@ class TimetableUnassignView(PanelAccessMixin, EventContextMixin, View):
         uow = self.request.di.uow
         try:
             TimetableService(uow).unassign_session(
-                session_pk, event_pk=current_event.pk, user_pk=self.request.user.pk
+                session_pk=session_pk,
+                event_pk=current_event.pk,
+                user_pk=self.request.user.pk,
             )
         except NotFoundError:
             return HttpResponse(status=422)
@@ -594,7 +573,7 @@ class TimetableRevertView(PanelAccessMixin, EventContextMixin, View):
         uow = self.request.di.uow
         try:
             TimetableService(uow).revert_change(
-                log_pk, event_pk=current_event.pk, user_pk=self.request.user.pk
+                log_pk=log_pk, event_pk=current_event.pk, user_pk=self.request.user.pk
             )
         except ValueError, NotFoundError:
             return HttpResponse(status=422)
