+++ src/ludamus/templates/components/tab-shell.html
@@ -1,4 +1,4 @@
-<div class="tab-shell flex flex-1 flex-col overflow-hidden rounded-2xl border border-border">
+<div class="tab-shell flex min-h-0 flex-1 flex-col overflow-hidden rounded-2xl border border-border">
     {% include "components/tab-shell-bar.html" with tabs_partial=tabs_partial %}
     {% if content|default:"" %}
         {{ content }}
