+++ src/ludamus/templates/panel/discounts/list.html
@@ -78,13 +78,12 @@
                                     {% tessera_button t_remove variant="danger" size="sm" icon="trash" full_width_mobile=False %}
                                 </form>
                             {% else %}
-                                <a href="?assign={{ row.facilitator.pk }}"
-                                   class="btn btn-secondary px-3 py-1.5 text-xs gap-1.5 rounded-xl"
-                                   aria-controls="discount-assign-modal-{{ row.facilitator.pk }}"
-                                   aria-haspopup="dialog">
-                                    {% icon "plus" class="w-3.5 h-3.5" %}
-                                    {% translate "Assign" %}
-                                </a>
+                                {% translate "Assign" as t_assign %}
+                                {% with facilitator_pk=row.facilitator.pk|stringformat:"s" %}
+                                    {% with assign_href="?assign="|add:facilitator_pk modal_id="discount-assign-modal-"|add:facilitator_pk %}
+                                        {% tessera_button t_assign href=assign_href variant="secondary" size="sm" icon="plus" full_width_mobile=False aria_controls=modal_id aria_haspopup="dialog" %}
+                                    {% endwith %}
+                                {% endwith %}
                             {% endif %}
                         </td>
                     </tr>
@@ -105,9 +104,12 @@
                         <div class="mt-4">{% tessera_form assignment.form %}</div>
                     </div>
                     <div class="flex justify-end gap-2 border-t border-border px-6 py-3">
-                        <button type="button"
-                                class="btn btn-secondary"
-                                data-modal-close="discount-assign-modal-{{ assignment.facilitator.pk }}">{% translate "Cancel" %}</button>
+                        {% with facilitator_pk=assignment.facilitator.pk|stringformat:"s" %}
+                            {% with modal_id="discount-assign-modal-"|add:facilitator_pk %}
+                                {% translate "Cancel" as t_cancel %}
+                                {% tessera_button t_cancel button_type="button" variant="secondary" full_width_mobile=False data_modal_close=modal_id %}
+                            {% endwith %}
+                        {% endwith %}
                         {% translate "Assign discount" as t_assign_discount %}
                         {% tessera_button t_assign_discount %}
                     </div>
