+++ src/ludamus/templates/panel/parts/sidebar-icon.html
@@ -1,5 +1,4 @@
-<span {% if icon_id|default:"" %}id="{{ icon_id }}"{% endif %}
-      class="block rotate-180">
+<span class="block rotate-180">
     <svg xmlns="http://www.w3.org/2000/svg"
          width="24"
          height="24"
