+++ src/ludamus/templates/panel/time-slots.html
@@ -121,10 +121,11 @@
                             {% endif %}
                         </div>
                         <div class="px-4 pb-4">
-                            <a href="?create={{ day|date:'Y-m-d' }}"
+                            <a href="?create=1&date={{ day|date:'Y-m-d' }}"
                                class="w-full px-3 py-2 text-sm border border-dashed border-border text-foreground-secondary rounded-lg hover:bg-bg-tertiary transition flex items-center justify-center"
-                               aria-controls="time-slot-create-modal-{{ day|date:'Ymd' }}"
-                               aria-haspopup="dialog">
+                               aria-controls="time-slot-create-modal"
+                               aria-haspopup="dialog"
+                               data-modal-reload>
                                 {% icon "plus" class="w-4 h-4 mr-1" %}
                                 {% translate "Add" %}
                             </a>
@@ -185,9 +186,6 @@
     {% end_tab_shell_body %}
 {% end_tab_shell %}
 {% include "panel/_time-slot-create-modal.html" with modal_id="time-slot-create-modal" form=create_form current_event=current_event csrf_token=csrf_token only %}
-{% for dated_create_form in dated_create_forms %}
-    {% include "panel/_time-slot-create-modal.html" with modal_id=dated_create_form.modal_id form=dated_create_form.form current_event=current_event csrf_token=csrf_token only %}
-{% endfor %}
 {% else %}
 <div class="bg-bg-secondary rounded-2xl shadow-(--shadow-card) border border-border p-8 text-center">
     <p class="text-foreground-muted">{% translate "No events available. Create an event to get started." %}</p>
